gpu_server/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.4
exceptiongroup==1.2.0
filelock==3.13.1
Flask==3.0.2
frozenlist==1.4.1
fsspec==2024.2.0
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.11
langchain-community==0.0.27
langchain-core==0.1.30
langchain-text-splitters==0.0.1
langsmith==0.1.22
llvmlite==0.42.0
MarkupSafe==2.1.5
marshmallow==3.21.1
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
openai-whisper==20231117
orjson==3.9.15
packaging==23.2
pillow==10.2.0
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
pydub==0.25.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
scipy==1.12.0
six==1.16.0
sniffio==1.3.1
sounddevice==0.4.6
SQLAlchemy==2.0.28
sympy==1.12
tenacity==8.2.3
tiktoken==0.6.0
torch==2.2.1
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
websockets==12.0
Werkzeug==3.0.1
yarl==1.9.4
